backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
aiosqlite
alembic
celery
redis
moviepy
langchain
langchain-openai
openai
python-docx
PyPDF2
httpx
python-multipart
pydantic
pydantic-settings
python-dotenv
aiofiles
Pillow
numpy
edge-tts
gevent
google-genai
